Move `QueryGraphRewriter` into its own module
